% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{d2}
\alias{d2}
\title{Example data from HW 2}
\format{An object of class \code{character} of length 2.}
\usage{
d2
}
\description{
Example data from HW 2
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q2p1.rda}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
