% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\docType{data}
\name{d3}
\alias{d3}
\title{Example data from HW 2}
\format{An object of class \code{character} of length 1.}
\usage{
d3
}
\description{
Example data from HW 2
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q2p3.rda}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
