% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangZheToolsRfunctions.R
\name{func8}
\alias{func8}
\title{x^T* (A^(− 1))* x}
\usage{
func8(a, x)
}
\arguments{
\item{d2}{data.frame}
}
\value{
scalar
}
\description{
given a numeric matrix A and a numeric vector x, calculates x^T* (A^(− 1))* x
}
\examples{
data(d2)
func8(a,x)


}
