% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangZheToolsRfunctions.R
\name{\%func9\%}
\alias{\%func9\%}
\title{x^T* (A^(− 1))* x}
\usage{
a \%func9\% x
}
\arguments{
\item{d2}{data.frame}
}
\value{
scalar
}
\description{
given a numeric matrix A and a numeric vector x, calculates x^T* (A^(− 1))* x
}
\examples{
data(d2)
a\%func9\%x


}
