% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangZheToolsmymarginals.R,
%   R/WangZheToolsmyplot.R
\name{plotMyData}
\alias{plotMyData}
\alias{plotMyData}
\title{Wrapper function for ggplot2 for data d}
\usage{
plotMyData(x)

plotMyData(x)
}
\arguments{
\item{x}{data.frame}

\item{x}{data.frame}
}
\value{
ggplot2

ggplot2
}
\description{
Computes the mean, variance and sd of a vector

Computes the mean, variance and sd of a vector
}
\examples{
data(d)
plotMyData(d)
data(d)
plotMyData(d)
}
